providers/supermicro: define a bmcQueryor interface and implement for x11, x12
based on feedback recieved this approach made more sense
