+++ providers/supermicro/floppy.go
@@ -24,7 +24,7 @@ func (c *Client) floppyImageMounted(ctx context.Context) (bool, error) {
 		return false, err
 	}
 
-	inserted, err := c.redfish.InsertedVirtualMedia(ctx)
+	inserted, err := c.serviceClient.redfish.InsertedVirtualMedia(ctx)
 	if err != nil {
 		return false, err
 	}
@@ -60,7 +60,7 @@ func (c *Client) MountFloppyImage(ctx context.Context, image io.Reader) error {
 		},
 		{
 			name: "csrf-token",
-			data: bytes.NewBufferString(c.csrfToken),
+			data: bytes.NewBufferString(c.serviceClient.csrfToken),
 		},
 	}
 
@@ -103,7 +103,7 @@ func (c *Client) MountFloppyImage(ctx context.Context, image io.Reader) error {
 	}
 	payloadWriter.Close()
 
-	resp, statusCode, err := c.query(
+	resp, statusCode, err := c.serviceClient.query(
 		ctx,
 		"cgi/uimapin.cgi",
 		http.MethodPost,
@@ -133,7 +133,7 @@ func (c *Client) UnmountFloppyImage(ctx context.Context) error {
 		return nil
 	}
 
-	resp, statusCode, err := c.query(
+	resp, statusCode, err := c.serviceClient.query(
 		ctx,
 		"cgi/uimapout.cgi",
 		http.MethodPost,
